pos: move wallet-side staking into libbitcoin_wallet
Second step of the RED-46 decoupling. The bodies that need CWallet move out
of libbitcoin_server into a new src/wallet/staking.{h,cpp}: CreateCoinStake,
FinalizeCoinStakeReward, GetStakeWeight over a wallet, and SignBlock. The file
also carries the only implementation of the interfaces::StakingWallet and
interfaces::StakingSupport seam added in the previous commit, and registers it
on NodeContext::staking_support from WalletInit::Construct(), so it stays null
in a --disable-wallet build.

What stays in the server library:

 - CheckBlockSignature. It is consensus, called from ContextualCheckBlock, and
   is wallet-free.
 - The chain-side coin-age weight loop in pos/stake.cpp, which reads block
   headers through the transaction index. It is retyped from
   std::set<CInputCoin> to std::vector<interfaces::StakeCoin> so pos/stake.h
   no longer needs wallet/wallet.h, and node/interfaces.cpp converts at the
   call site until that method goes away.

miner.cpp, rpc/mining.cpp and test/util/setup_common.cpp still call the moved
functions by name and now include wallet/staking.h for their declarations.
Later phases route them through the seam instead, after which the declarations
become internal to wallet/staking.cpp. The wallet-enabled build is unchanged;
the no-wallet build does not link yet.

Note that CreateCoinStake's kernel search still reaches server symbols
(CheckStakeKernelHash, g_txindex, OpenBlockFile, DeploymentActiveAfter), so
the wallet library is not free of server references on this path. This is
recorded as a known exception in the design doc, along with the fallbacks if
either link order rejects it.
